requirements.txt
fastapi==0.111.0
PyJWT>=2.10.1,<3.0.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.1
pydantic==2.7.0
supabase==2.4.3
anthropic==0.28.0
playwright==1.44.0
httpx==0.27.0
python-multipart==0.0.9
jinja2==3.1.4
aiofiles==23.2.1
